requirements.txt
schedule
pyserial
influxdb-client
python-dotenv
